Refactoring to make it easier to add XeSS
