+++ RenderCore/render/backend/render_backend.hpp
@@ -6,8 +6,8 @@
 #include <VkBootstrap.h>
 #include <tracy/TracyVulkan.hpp>
 
-#include "hit_group_builder.hpp"
-#include "render/noise_texture.hpp"
+#include "streamline_adapter.hpp"
+#include "render/backend/hit_group_builder.hpp"
 #include "render/backend/descriptor_set_allocator.hpp"
 #include "render/backend/render_graph.hpp"
 #include "render/backend/resource_access_synchronizer.hpp"
@@ -18,7 +18,6 @@
 #include "render/backend/resource_upload_queue.hpp"
 #include "render/backend/constants.hpp"
 
-class StreamlineAdapter;
 class BlasBuildQueue;
 class PipelineCache;
 /**
@@ -100,11 +99,6 @@ public:
     uint32_t get_current_gpu_frame() const;
 
     /**
-     * Updates internal state with the beginning of the simulation. Useful for things like Nvidia Reflex
-     */
-    void mark_simulation_begin() const;
-
-    /**
      * Begins the frame
      *
      * Waits for the GPU to finish with this frame, does some beginning-of-frame setup, is generally cool
@@ -145,10 +139,6 @@ public:
      * Callers should make no effort to save these descriptors
      */
     DescriptorSetAllocator& get_transient_descriptor_allocator();
-
-#if SAH_USE_STREAMLINE
-    StreamlineAdapter* get_streamline() const;
-#endif
 
     CommandBuffer create_graphics_command_buffer(const std::string& name);
 
@@ -200,10 +190,6 @@ public:
 private:
     static inline std::unique_ptr<RenderBackend> g_render_backend = nullptr;
 
-#if SAH_USE_STREAMLINE
-    std::unique_ptr<StreamlineAdapter> streamline;
-#endif
-
     bool is_first_frame = true;
 
     uint32_t cur_frame_idx = 0;
@@ -229,6 +215,8 @@ private:
 
     VkQueue transfer_queue;
     uint32_t transfer_queue_family_index;
+
+    StreamlineAdapter streamline;
 
     std::unique_ptr<ResourceAllocator> allocator;
