+++ RenderCore/render/backend/resource_access_synchronizer.hpp
@@ -2,7 +2,6 @@
 #include <vector>
 #include <volk.h>
 
-#include "acceleration_structure_usage_token.hpp"
 #include "render/backend/buffer_usage_token.hpp"
 #include "render/backend/texture_usage_token.hpp"
 #include "render/backend/handles.hpp"
