+++ RenderCore/render/backend/streamline_adapter.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#if SAH_USE_STREAMLINE
+#include <vulkan/vulkan_core.h>
+
+class StreamlineAdapter
+{
+public:
+    StreamlineAdapter();
+
+    ~StreamlineAdapter();
+
+    PFN_vkGetInstanceProcAddr try_load_interposer();
+};
+#endif
