+++ RenderCore/render/phase/ambient_occlusion_phase.cpp
@@ -7,6 +7,7 @@
 
 #include "console/cvars.hpp"
 #include "core/string_conversion.hpp"
+#include "render/noise_texture.hpp"
 #include "render/render_scene.hpp"
 #include "render/scene_view.hpp"
 #include "render/backend/pipeline_cache.hpp"
