+++ RenderCore/render/upscaling/fsr3.cpp
@@ -4,8 +4,8 @@
 
 #include <utf8.h>
 
-#include "scene_view.hpp"
-#include "backend/render_backend.hpp"
+#include "render/scene_view.hpp"
+#include "render/backend/render_backend.hpp"
 #include "console/cvars.hpp"
 #include "core/system_interface.hpp"
 #include "core/string_conversion.hpp"
@@ -18,10 +18,12 @@ static std::string to_string(FfxApiUpscaleQualityMode quality_mode);
 
 static std::shared_ptr<spdlog::logger> logger;
 
-FidelityFSSuperResolution3::FidelityFSSuperResolution3(RenderBackend& backend) {
+FidelityFSSuperResolution3::FidelityFSSuperResolution3() {
     if(logger == nullptr) {
         logger = SystemInterface::get().get_logger("FidelityFSSuperResolution3");
     }
+
+    auto& backend = RenderBackend::get();
 
     backend_desc.vkDevice = backend.get_device();
     backend_desc.vkPhysicalDevice = backend.get_physical_device();
@@ -35,7 +37,7 @@ FidelityFSSuperResolution3::~FidelityFSSuperResolution3() {
     }
 }
 
-void FidelityFSSuperResolution3::initialize(const glm::uvec2& output_resolution_in) {
+void FidelityFSSuperResolution3::initialize(const glm::uvec2 output_resolution_in, const uint32_t frame_number) {
     output_resolution = output_resolution_in;
     glm::uvec2 new_render_resolution;
     ffx::QueryDescUpscaleGetRenderResolutionFromQualityMode query = {};
@@ -65,15 +67,15 @@ void FidelityFSSuperResolution3::initialize(const glm::uvec2& output_resolution_
     if(!has_context) {
         ffx::CreateContextDescUpscale create_upscaling;
         create_upscaling.flags = FFX_UPSCALE_ENABLE_HIGH_DYNAMIC_RANGE |
-            FFX_UPSCALE_ENABLE_AUTO_EXPOSURE | 
-            FFX_UPSCALE_ENABLE_DEPTH_INFINITE | 
+            FFX_UPSCALE_ENABLE_AUTO_EXPOSURE |
+            FFX_UPSCALE_ENABLE_DEPTH_INFINITE |
             FFX_UPSCALE_ENABLE_DEBUG_CHECKING;
         create_upscaling.maxRenderSize = {.width = optimal_render_resolution.x, .height = optimal_render_resolution.y};
         create_upscaling.maxUpscaleSize = {.width = output_resolution.x, .height = output_resolution.y};
         create_upscaling.fpMessage = +[](const uint32_t type, const wchar_t* c_message) {
             const auto wmessage = std::u16string_view{reinterpret_cast<const char16_t*>(c_message)};
             const auto message = utf8::utf16tou8(wmessage);
-            const auto skill_issue = std::string_view{ reinterpret_cast<const char*>(message.c_str()), message.size() };
+            const auto skill_issue = std::string_view{reinterpret_cast<const char*>(message.c_str()), message.size()};
             if(type == FFX_API_MESSAGE_TYPE_WARNING) {
                 logger->warn("{}", skill_issue);
             } else if(type == FFX_API_MESSAGE_TYPE_ERROR) {
@@ -112,12 +114,12 @@ void FidelityFSSuperResolution3::initialize(const glm::uvec2& output_resolution_
 }
 
 void FidelityFSSuperResolution3::set_constants(const SceneView& scene_transform, const glm::uvec2 render_resolution) {
-    const auto f_render_resolution = glm::vec2{ render_resolution };
-    const auto scaled_jitter = scene_transform.get_jitter() * f_render_resolution * glm::vec2{ 0.5f, -0.5f };
+    const auto f_render_resolution = glm::vec2{render_resolution};
+    const auto scaled_jitter = scene_transform.get_jitter();// *f_render_resolution* glm::vec2{ 0.5f, -0.5f };
     dispatch_desc.jitterOffset = {-scaled_jitter.x, -scaled_jitter.y};
-    dispatch_desc.motionVectorScale = { 1.f, 1.f };
+    dispatch_desc.motionVectorScale = {1.f, 1.f};
     dispatch_desc.renderSize = {render_resolution.x, render_resolution.y};
-    dispatch_desc.upscaleSize = { output_resolution.x, output_resolution.y };
+    dispatch_desc.upscaleSize = {output_resolution.x, output_resolution.y};
     dispatch_desc.frameTimeDelta = 7.5f; // Hardcoded, please fix
     dispatch_desc.preExposure = 1.f;
     dispatch_desc.cameraNear = scene_transform.get_near();
@@ -126,48 +128,83 @@ void FidelityFSSuperResolution3::set_constants(const SceneView& scene_transform,
     dispatch_desc.viewSpaceToMetersFactor = 1.f;
 }
 
-void FidelityFSSuperResolution3::dispatch(
-    const CommandBuffer& commands, const TextureHandle color_in, const TextureHandle color_out,
+void FidelityFSSuperResolution3::evaluate(
+    RenderGraph& graph, const TextureHandle color_in, const TextureHandle color_out,
     const TextureHandle depth_in, const TextureHandle motion_vectors_in
 ) {
-    const auto color_in_res = FfxApiResource{
-        .resource = color_in->image,
-        .description = ffxApiGetImageResourceDescriptionVK(color_in->image, color_in->create_info, 0),
-        .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
-    };
-    const auto color_out_res = FfxApiResource{
-        .resource = color_out->image,
-        .description = ffxApiGetImageResourceDescriptionVK(color_out->image, color_out->create_info, 0),
-        .state = FFX_API_RESOURCE_STATE_UNORDERED_ACCESS
-    };
-    const auto depth_in_res = FfxApiResource{
-        .resource = depth_in->image,
-        .description = ffxApiGetImageResourceDescriptionVK(depth_in->image, depth_in->create_info, 0),
-        .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
-    };
-    const auto motion_vectors_in_res = FfxApiResource{
-        .resource = motion_vectors_in->image,
-        .description = ffxApiGetImageResourceDescriptionVK(motion_vectors_in->image, motion_vectors_in->create_info, 0),
-        .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
-    };
+    graph.add_pass(
+        {
+            .name = "fsr3",
+            .textures = {
+                {
+                    .texture = color_in,
+                    .stage = VK_PIPELINE_STAGE_2_ALL_COMMANDS_BIT,
+                    .access = VK_ACCESS_2_SHADER_READ_BIT,
+                    .layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL
+                },
+                {
+                    .texture = color_out,
+                    .stage = VK_PIPELINE_STAGE_2_ALL_COMMANDS_BIT,
+                    .access = VK_ACCESS_2_SHADER_READ_BIT | VK_ACCESS_2_SHADER_WRITE_BIT,
+                    .layout = VK_IMAGE_LAYOUT_GENERAL
+                },
+                {
+                    .texture = depth_in,
+                    .stage = VK_PIPELINE_STAGE_2_ALL_COMMANDS_BIT,
+                    .access = VK_ACCESS_2_SHADER_READ_BIT,
+                    .layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL
+                },
+                {
+                    .texture = motion_vectors_in,
+                    .stage = VK_PIPELINE_STAGE_2_ALL_COMMANDS_BIT,
+                    .access = VK_ACCESS_2_SHADER_READ_BIT,
+                    .layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL
+                },
+            },
+            .execute = [&](CommandBuffer& commands) {
+                const auto color_in_res = FfxApiResource{
+                    .resource = color_in->image,
+                    .description = ffxApiGetImageResourceDescriptionVK(color_in->image, color_in->create_info, 0),
+                    .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
+                };
+                const auto color_out_res = FfxApiResource{
+                    .resource = color_out->image,
+                    .description = ffxApiGetImageResourceDescriptionVK(color_out->image, color_out->create_info, 0),
+                    .state = FFX_API_RESOURCE_STATE_UNORDERED_ACCESS
+                };
+                const auto depth_in_res = FfxApiResource{
+                    .resource = depth_in->image,
+                    .description = ffxApiGetImageResourceDescriptionVK(depth_in->image, depth_in->create_info, 0),
+                    .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
+                };
+                const auto motion_vectors_in_res = FfxApiResource{
+                    .resource = motion_vectors_in->image,
+                    .description = ffxApiGetImageResourceDescriptionVK(
+                        motion_vectors_in->image,
+                        motion_vectors_in->create_info,
+                        0),
+                    .state = FFX_API_RESOURCE_STATE_COMPUTE_READ
+                };
 
-    auto local_dispatch_desc = dispatch_desc;
-    local_dispatch_desc.commandList = commands.get_vk_commands();
-    local_dispatch_desc.color = color_in_res;
-    local_dispatch_desc.depth = depth_in_res;
-    local_dispatch_desc.motionVectors = motion_vectors_in_res;
-    local_dispatch_desc.output = color_out_res;
-    // local_dispatch_desc.flags = FFX_UPSCALE_FLAG_DRAW_DEBUG_VIEW;
+                auto local_dispatch_desc = dispatch_desc;
+                local_dispatch_desc.commandList = commands.get_vk_commands();
+                local_dispatch_desc.color = color_in_res;
+                local_dispatch_desc.depth = depth_in_res;
+                local_dispatch_desc.motionVectors = motion_vectors_in_res;
+                local_dispatch_desc.output = color_out_res;
+                // local_dispatch_desc.flags = FFX_UPSCALE_FLAG_DRAW_DEBUG_VIEW;
 
-    ffx::Dispatch(upscaling_context, local_dispatch_desc);
+                ffx::Dispatch(upscaling_context, local_dispatch_desc);
+            }
+        });
 }
 
 glm::uvec2 FidelityFSSuperResolution3::get_optimal_render_resolution() const {
     return optimal_render_resolution;
 }
 
-glm::vec2 FidelityFSSuperResolution3::get_jitter() const {
-    return jitter * glm::vec2{ 2.f, -2.f } / glm::vec2{ optimal_render_resolution };
+glm::vec2 FidelityFSSuperResolution3::get_jitter() {
+    return jitter;
 }
 
 std::string to_string(const FfxApiUpscaleQualityMode quality_mode) {
