+++ RenderCore/shaders/motion_vectors/motion_vectors_opaque.frag.slang
@@ -19,7 +19,7 @@ float2 main(VertexOutput vertex) {
     const float2 previous_uv_location = previous_clipspace_location * float2(0.5, 0.5) + 0.5;
     const float2 previous_pixel_location = previous_uv_location * camera_data.render_resolution;
 
-    const float2 pixel_motion_vectors = previous_pixel_location - camera_data.previous_jitter - vertex.position.xy - camera_data.jitter;
+    const float2 pixel_motion_vectors = previous_pixel_location - vertex.position.xy;
 
     return pixel_motion_vectors;
 }
